Redesigned option-setting mechanism (yes, again).
Instead of many set_option() overload methods on Connection (plus
parallel ones on DBDriver) taking an option enum value and an optional
argument, it now takes a pointer to an abstract Option object.  There
is a subclass for each option we support, and each one deals with
options, and implements a set() method which DBDriver calls to apply
the option.

This leaves Connection with one trivial set_option() method, DBDriver
with a smaller set of simple set_option() overloads, and a passel of
cookie-cutter Option subclass implementations.  SLOC went up as a result
of the latter even though a lot of large case statements got the axe,
but because it's cookie-cutter code, it imposes no mental overhead.

Biggest benefit is that it does away with the "bad argument type"
error.  It just can't happen now.